+++ lib/agilibox/engine.rb
@@ -11,4 +11,5 @@ module Agilibox
   end
 end
 
+require_relative "engine_file"
 require_relative "core_and_rails_ext"