+++ app/filters/agilibox/small_data/filter_strategy_by_date_begin.rb
@@ -1,6 +1,6 @@
 class Agilibox::SmallData::FilterStrategyByDateBegin < ::Agilibox::SmallData::FilterStrategyByKeyValue
   def apply(query, value)
-    value = Time.parse(value).beginning_of_day
+    value = Time.zone.parse(value).beginning_of_day
     query.where("#{key} >= ?", value)
   end
 end