+++ app/filters/agilibox/small_data/filter_strategy_by_date_end.rb
@@ -1,6 +1,6 @@
 class Agilibox::SmallData::FilterStrategyByDateEnd < ::Agilibox::SmallData::FilterStrategyByKeyValue
   def apply(query, value)
-    value = Time.parse(value).end_of_day
+    value = Time.zone.parse(value).end_of_day
     query.where("#{key} <= ?", value)
   end
 end