+++ spec/helpers/agilibox/sorting_helper_spec.rb
@@ -1,11 +1,13 @@
-require 'rails_helper'
+require "rails_helper"
 
 describe Agilibox::SortingHelper, type: :helper do
   describe "#sortable_column" do
-    let(:params) {{
-      :controller => "sort",
-      :action     => "sort",
-    }}
+    let(:params) {
+      {
+        :controller => "sort",
+        :action     => "sort",
+      }
+    }
 
     it "current sort is nil" do
       link = sortable_column("text", :col)
@@ -13,25 +15,25 @@ describe Agilibox::SortingHelper, type: :helper do
     end
 
     it "current sort is col" do
-      params.merge!(sort: "col")
+      params[:sort] = "col"
       link = sortable_column("text", :col)
       expect(link).to eq %(<a class="sort asc" href="/?sort=-col">text ↓</a>)
     end
 
     it "current sort is -col" do
-      params.merge!(sort: "-col")
+      params[:sort] = "-col"
       link = sortable_column("text", :col)
       expect(link).to eq %(<a class="sort desc" href="/?sort=col">text ↑</a>)
     end
 
     it "current sort is other" do
-      params.merge!(sort: "other")
+      params[:sort] = "other"
       link = sortable_column("text", :col)
       expect(link).to eq %(<a class="sort" href="/?sort=col">text</a>)
     end
 
     it "current sort is -other" do
-      params.merge!(sort: "-other")
+      params[:sort] = "-other"
       link = sortable_column("text", :col)
       expect(link).to eq %(<a class="sort" href="/?sort=col">text</a>)
     end
@@ -39,7 +41,7 @@ describe Agilibox::SortingHelper, type: :helper do
     it "should raise on invalid column type" do
       # old handles_sortable_columns syntax
       expect {
-        sortable_column("text", {column: "col"})
+        sortable_column("text", column: "col")
       }.to raise_error ArgumentError
     end
   end # describe "#sortable_column"