+++ spec/dummy/app/models/dummy_model.rb
@@ -1,5 +1,6 @@
 class DummyModel < ApplicationRecord
   include Agilibox::Search
+  include Agilibox::TimestampHelpers
 
   belongs_to :asso, polymorphic: true, required: false