+++ spec/dummy/app/controllers/dummy_controller.rb
@@ -1,0 +1,6 @@
+class DummyController < ApplicationController
+  include Agilibox::ApiControllerConcern
+
+  def show
+  end
+end